LED drivers: rename "simple" to "mono"
