+++ drivers/led/issi/is31fl3733-mono.c
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "is31fl3733-simple.h"
+#include "is31fl3733-mono.h"
 #include <string.h>
 #include "i2c_master.h"
 #include "wait.h"
