+++ drivers/led/issi/is31fl3736-mono.c
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "is31fl3736-simple.h"
+#include "is31fl3736-mono.h"
 #include <string.h>
 #include "i2c_master.h"
 #include "wait.h"
