+++ drivers/led/issi/is31fl3745-mono.c
@@ -1,4 +1,4 @@
-#include "is31fl3745-simple.h"
+#include "is31fl3745-mono.h"
 #include <string.h>
 #include "i2c_master.h"
 #include "wait.h"
