+++ keyboards/fallacy/rules.mk
@@ -13,5 +13,5 @@ AUDIO_ENABLE = no           # Audio output
 
 # project specific files
 SRC += indicators.c \
-       drivers/led/issi/is31fl3731-simple.c
+       drivers/led/issi/is31fl3731-mono.c
 I2C_DRIVER_REQUIRED = yes
